create UserInterface and App classes
