+++ src/index.js
@@ -1,4 +1,8 @@
+const App = require('./app')
 const ReadingList = require('./reading-list')
+const UserInterface = require('./user-interface')
 
 const readingList = new ReadingList()
-readingList.open()
+const userInterface = new UserInterface()
+const app = new App(readingList, userInterface)
+app.start()
